\name{pracopt-package}
\alias{pracopt-package}
\alias{pracopt}
\docType{package}
\title{
  Optimization Test Routines
}
\description{
         pracopt provides test functions for local and global
         optimization, both constrained and unconstrained.
}
\details{
\tabular{ll}{
Package: \tab pracopt\cr
Type: \tab Package\cr
Version: \tab 0.1.5\cr
Date: \tab 2011-07-27\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Hans W Borchers

Maintainer: HwB  <hwborchers@googlemail.com>
}
\references{
  A. Antoniou and W.-S. Lu (2007). Practical Optimization: Algorithms and 
  Engineering Applications. Springer Science+Business Media, New York.
}
\keyword{ package }
\seealso{
  The R-Forge project ``Optimization and solving packages''.
  \url{https://r-forge.r-project.org/projects/optimizer/}
}
\examples{
# See the functions in this package.
}
